+++ src/main/java/online/kingdomkeys/kingdomkeys/magic/MagicFire.java
@@ -18,7 +18,8 @@ public class MagicFire extends Magic {
 
 	@Override
 	public void magicUse(LivingEntity player, Player caster, int level, float fullMPBlastMult, LivingEntity lockOnEntity) {
-		float dmgMult = getDamageMult(level) + PlayerData.get(caster).getNumberOfAbilitiesEquipped(Strings.fireBoost) * 0.2F;
+		float dmgMult = getRealDamageMult(level, caster) + PlayerData.get(caster).getNumberOfAbilitiesEquipped(Strings.fireBoost) * 0.2F;
+		System.out.println(getRealDamageMult(level,caster));
 		dmgMult *= fullMPBlastMult;
 		lockOnEntity = getMagicLockOn(level) ? lockOnEntity : null;